# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MBIO PAY — Crypto to Mobile Money Remittance API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface QuoteResponse {
  usdtAmount: number;
  payoutUGX: number;
  usdtRate: number;
  fee: number;
}

export interface WalletAddressResponse {
  address: string;
  network: string;
}

export interface WalletBalanceResponse {
  usdtBalance: number;
  ugxEquivalent: number;
  usdtRate: number;
  address: string;
}

export interface StatsResponse {
  totalOrders: number;
  completedOrders: number;
  totalUgxPaidOut: number;
  totalUsdtReceived: number;
  pendingOrders: number;
}

export type CreateOrderRequestNetwork =
  (typeof CreateOrderRequestNetwork)[keyof typeof CreateOrderRequestNetwork];

export const CreateOrderRequestNetwork = {
  MTN: "MTN",
  Airtel: "Airtel",
} as const;

export interface CreateOrderRequest {
  phone: string;
  network: CreateOrderRequestNetwork;
  /** USDT amount the user expects to send */
  expectedUsdt: number;
}

export interface CreateOrderResponse {
  orderId: number;
  address: string;
  message: string;
  payoutUGX: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  waiting: "waiting",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface Order {
  id: number;
  phone: string;
  network: string;
  amount?: number | null;
  ugxAmount?: number | null;
  status: OrderStatus;
  txid?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetQuoteParams = {
  /**
   * Amount in USDT
   */
  amount: number;
};