# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MBIO PAY — Crypto to Mobile Money Remittance API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: wallet
  - name: orders
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /quote:
    get:
      operationId: getQuote
      tags: [orders]
      summary: Get UGX payout quote for a USDT amount
      parameters:
        - name: amount
          in: query
          required: true
          schema:
            type: number
          description: Amount in USDT
      responses:
        "200":
          description: Quote details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteResponse"

  /wallet/address:
    get:
      operationId: getWalletAddress
      tags: [wallet]
      summary: Get deposit wallet address
      responses:
        "200":
          description: Wallet address and network info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletAddressResponse"

  /wallet/balance:
    get:
      operationId: getWalletBalance
      tags: [wallet]
      summary: Get USDT and UGX balance
      responses:
        "200":
          description: Balance details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalanceResponse"

  /stats:
    get:
      operationId: getStats
      tags: [wallet]
      summary: Get dashboard stats
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResponse"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create remittance order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOrderResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/recent:
    get:
      operationId: getRecentOrders
      tags: [orders]
      summary: Get recent orders
      responses:
        "200":
          description: List of recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    QuoteResponse:
      type: object
      properties:
        usdtAmount:
          type: number
        payoutUGX:
          type: number
        usdtRate:
          type: number
        fee:
          type: number
      required: [usdtAmount, payoutUGX, usdtRate, fee]

    WalletAddressResponse:
      type: object
      properties:
        address:
          type: string
        network:
          type: string
      required: [address, network]

    WalletBalanceResponse:
      type: object
      properties:
        usdtBalance:
          type: number
        ugxEquivalent:
          type: number
        usdtRate:
          type: number
        address:
          type: string
      required: [usdtBalance, ugxEquivalent, usdtRate, address]

    StatsResponse:
      type: object
      properties:
        totalOrders:
          type: integer
        completedOrders:
          type: integer
        totalUgxPaidOut:
          type: number
        totalUsdtReceived:
          type: number
        pendingOrders:
          type: integer
      required: [totalOrders, completedOrders, totalUgxPaidOut, totalUsdtReceived, pendingOrders]

    CreateOrderRequest:
      type: object
      properties:
        phone:
          type: string
        network:
          type: string
          enum: [MTN, Airtel]
        expectedUsdt:
          type: number
          description: USDT amount the user expects to send
      required: [phone, network, expectedUsdt]

    CreateOrderResponse:
      type: object
      properties:
        orderId:
          type: integer
        address:
          type: string
        message:
          type: string
        payoutUGX:
          type: number
      required: [orderId, address, message, payoutUGX]

    Order:
      type: object
      properties:
        id:
          type: integer
        phone:
          type: string
        network:
          type: string
        amount:
          type: number
          nullable: true
        ugxAmount:
          type: number
          nullable: true
        status:
          type: string
          enum: [waiting, processing, completed, failed]
        txid:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, phone, network, status, createdAt, updatedAt]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]