# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MBIO PAY — Crypto to Mobile Money Remittance API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderRequestNetwork } from "./createOrderRequestNetwork";

export interface CreateOrderRequest {
  phone: string;
  network: CreateOrderRequestNetwork;
  /** USDT amount the user expects to send */
  expectedUsdt: number;
}