# lib/api-zod/src/generated/types/quoteResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MBIO PAY — Crypto to Mobile Money Remittance API
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteResponse {
  usdtAmount: number;
  payoutUGX: number;
  usdtRate: number;
  fee: number;
}