# lib/api-zod/src/generated/types/walletAddressResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MBIO PAY — Crypto to Mobile Money Remittance API
 * OpenAPI spec version: 0.1.0
 */

export interface WalletAddressResponse {
  address: string;
  network: string;
}